fix: 회원탈퇴 시 s3 필수서류 삭제

See merge request s08-webmobile2-sub2/S08P12C107!87

+++ back/src/main/java/com/web/puppylink/service/VolunteerServiceImpl.java
@@ -294,16 +294,20 @@ public class VolunteerServiceImpl implements VolunteerService{
 				fileUrl = volunteer.getPassportURL();
 			}
 			
-			String fileKey= fileUrl.split("com/")[1]; 		
+			// fileURL없는 경우
+			if(fileUrl != null && !fileUrl.equals("")) {
+				String fileKey= fileUrl.split("com/")[1]; 		
 
-            final AmazonS3 s3 = AmazonS3ClientBuilder.standard().withRegion(Region).build();
+	            final AmazonS3 s3 = AmazonS3ClientBuilder.standard().withRegion(Region).build();
 
-            try {
-                s3.deleteObject(Bucket, fileKey);
-            } catch (AmazonServiceException e) {
-                System.err.println(e.getErrorMessage());
-                System.exit(1);
-            }
+	            try {
+	                s3.deleteObject(Bucket, fileKey);
+	            } catch (AmazonServiceException e) {
+	                System.err.println(e.getErrorMessage());
+	                System.exit(1);
+	            }
+			}
+			
 	    } catch (Exception e) {
 	       e.printStackTrace();
 	       throw new RuntimeException("s3 객체를 삭제하는데 실패했습니다.");
